fix(position): 修复职位浏览次数数据类型不一致问题
- 将Position实体中viewCount字段的数据类型统一为Long类型
- 修正PositionServiceImpl中viewCount的类型声明和初始化值
- 确保数据库中浏览次数字段的数据类型保持一致
- 避免因数据类型不匹配导致的潜在类型转换异常

+++ src/main/java/com/xiyuxian/positionmatch/service/impl/PositionServiceImpl.java
@@ -41,7 +41,7 @@ public class PositionServiceImpl extends ServiceImpl<PositionMapper, Position> i
 
         Position position = new Position();
         BeanUtil.copyProperties(positionAddRequest, position);
-        position.setViewCount(0);
+        position.setViewCount(0L);
         position.setApplyCount(0L);
         position.setCreateTime(new Date());
         position.setUpdateTime(new Date());
@@ -175,9 +175,9 @@ public class PositionServiceImpl extends ServiceImpl<PositionMapper, Position> i
             throw new BusinessException(ErrorCode.NOT_FOUND_ERROR, "职位不存在");
         }
 
-        Integer viewCount = position.getViewCount();
+        Long viewCount = position.getViewCount();
         if (viewCount == null) {
-            viewCount = 0;
+            viewCount = 0L;
         }
         position.setViewCount(viewCount + 1);
         return this.updateById(position);
